refactor: 定数のページ情報を整理
