+++ src/components/Navigation/Navigation.tsx
@@ -2,17 +2,17 @@ import { NavLink } from '@solidjs/router'
 import { Index } from 'solid-js'
 
 import styles from '@/components/Navigation/Navigation.module.scss'
-import { links } from '@/constants/index'
+import { PAGES_INFO } from '@/constants'
 
 const Navigation = () => {
   return (
     <nav class={styles.root}>
       <ul class={styles.container}>
-        <Index each={links}>
-          {(link) => (
+        <Index each={PAGES_INFO}>
+          {(page) => (
             <li class={styles.item}>
-              <NavLink href={link().URL} activeClass={styles.active} class={styles.link} end>
-                {link().name}
+              <NavLink href={page().path} activeClass={styles.active} class={styles.link} end>
+                {page().title.japanese}
               </NavLink>
             </li>
           )}
