+++ src/types/index.ts
@@ -1,0 +1,3 @@
+export type DeepReadonly<T> = {
+  readonly [P in keyof T]: DeepReadonly<T[P]>
+}
